fix steamcmd repeat

+++ internal/app/game_server_commands/install_server.go
@@ -560,7 +560,7 @@ func (in *installator) installFromSteam(
 			break
 		}
 
-		if !repeatableSteamCMDInstallResults.Contains(int8(result)) {
+		if !repeatableSteamCMDInstallResults.Contains(result) {
 			break
 		}
